+++ lib/sidekiq-spy/display/panels/header.rb
@@ -1,3 +1,6 @@
+require 'time'
+
+
 module SidekiqSpy
   module Display
     module Panels