+++ lib/sidekiq-spy/spy/dataspyable.rb
@@ -1,0 +1,17 @@
+module SidekiqSpy
+  module Spy
+    module Dataspyable
+      
+      include Enumerable
+      
+      def each(&block)
+        @data.each(&block)
+      end
+      
+      def [](key)
+        @data[key] || {}
+      end
+      
+    end
+  end
+end